// src/pages/Memoji/data/enums.ts
export enum ResultStatus {
  NONE = 'NONE',
  WIN = 'WIN',
  LOSE = 'LOSE',
}

export enum GameStatus {
  NONE = 'NONE',
  STOP = 'STOP',
  PLAY = 'PLAY',
  PAUSE = 'PAUSE',
}

export enum SetTypes {
  ANIMALS = 'ANIMALS',
  FINGERS = 'FINGERS',
  FLAGS = 'FLAGS',
}

export enum Theme {
  LIGHT = 'LIGHT',
  DARK = 'DARK',
  LEMON = 'LEMON',
  RAINBOW = 'RAINBOW',
}